remove ring-cljsjs; fixes #132
This adds routes for requesting resources from cljsjs, but we don't seem to be using this (maybe never did). My guess is that I experimented with using this to request cljsjs assets dynamically, but we're not even using cljsjs anymore (shadow-cljs instead), so this seems like a dead end. This particular dep is also causing problems with dependency resolution in leiningen (see #132); so, good ridance!