Make clojupyter an optional dep, not provided
As of tools.deps.alpha 0.9.745, when generating a pom it is possible to
mark a dependency as `<optional>true</optional>`. This is similar to
`<scope>provided</scope>` in that it indicates that the user of the
library must arrange for the dependency to be on the classpath. However,
the semantics of "optional" are more that this dependency is only
sometimes needed.

This patch replaces the various hacks that marked clojupyter as a
provided dependency with an indication that it is an optional
dependency. This should have the desired effect of letting tooling like
cljdocs load clojupyter, while still keeping clojupyter out of the
dependency trees of oz users, unless they explicitly add it.

I discovered this worked because I was having similar problems with
cljdocs in a different library, and ran across some threads that led to
oz.